types file into base
