+++ internal/auth0/connection_new/auth0_test.go
@@ -1,4 +1,4 @@
-package connection_new_test //nolint:all
+package connection_new_test //nolint:all temporarily until v0 connection resource removed
 
 import (
 	"fmt"
